feat: moved non startup IPC calls to their own space
